Remove defaultRepo and split up sync.ts
